Move test tasks to a separate module and conditionally import. Also fix exportchannel, which shold not be trying to use async task API yet.
