+++ contentcuration/contentcuration/tasks.py
@@ -11,7 +11,6 @@ from django.core.mail import EmailMessage
 from django.core.management import call_command
 from django.template.loader import render_to_string
 
-from contentcuration.celery import app
 from contentcuration.models import Channel
 from contentcuration.models import ContentNode
 from contentcuration.models import Task
@@ -20,6 +19,11 @@ from contentcuration.utils.csv_writer import write_channel_csv_file
 from contentcuration.utils.csv_writer import write_user_csv
 
 logger = get_task_logger(__name__)
+
+
+# if we're running tests, import our test tasks as well
+if settings.RUNNING_TESTS:
+    from .tasks_test import *
 
 # TODO: Try to get debugger working for celery workers
 # Attach Python Cloud Debugger
@@ -39,52 +43,8 @@ logger = get_task_logger(__name__)
 
 # runs the management command 'exportchannel' async through celery
 
-
-# Tasks with 'test' in their name are written specifically to test the Async Task API
-@app.task(bind=True, name='test_task')
-def test_task(self):
-    """
-    This is a mock task to be used ONLY for unit-testing various pieces of the
-    async task API.
-    :return: The number 42
-    """
-    logger.info("Request ID = {}".format(self.request.id))
-    assert Task.objects.filter(task_id=self.request.id).count() == 1
-    return 42
-
-
-@app.task(name='error_test_task')
-def error_test_task():
-    """
-    This is a mock task designed to test that we properly report errors to the client.
-    """
-    raise Exception("I'm sorry Dave, I'm afraid I can't do that.")
-
-
-@app.task(bind=True, name='progress_test_task')
-def progress_test_task(self):
-    """
-    This is a mock task to be used to test that we can update progress when tracking is enabled.
-    :return:
-    """
-    logger.info("Request ID = {}".format(self.request.id))
-    assert Task.objects.filter(task_id=self.request.id).count() == 1
-
-    self.update_state('PROGRESS', meta={'progress': 100})
-    return 42
-
-
-@app.task(name='non_async_test_task')
-def non_async_test_task():
-    """
-    This is a mock task used to test that creating a task without using create_async_task does not result
-    in a Task object being created or updated.
-    """
-    return 42
-
-
 @task(name='exportchannel_task')
-def exportchannel_task(channel_id, user_id, task_type, is_progress_tracking, metadata):
+def exportchannel_task(channel_id, user_id):
     call_command('exportchannel', channel_id, email=True, user_id=user_id)
 
 
@@ -125,11 +85,13 @@ def deletetree_task(tree_id):
     ContentNode.objects.filter(tree_id=tree_id).delete()
 
 
-type_mapping = {
-    'test': {'task': test_task, 'progress_tracking': False},
-    'error-test': {'task': error_test_task, 'progress_tracking': False},
-    'progress-test': {'task': progress_test_task, 'progress_tracking': True},
-}
+type_mapping = {}
+if settings.RUNNING_TESTS:
+    type_mapping.update({
+        'test': {'task': test_task, 'progress_tracking': False},
+        'error-test': {'task': error_test_task, 'progress_tracking': False},
+        'progress-test': {'task': progress_test_task, 'progress_tracking': True}
+    })
 
 
 def create_async_task(task_name, task_options, task_args=None):
