+++ contentcuration/contentcuration/views/base.py
@@ -344,13 +344,10 @@ def publish_channel(request):
     except KeyError:
         raise ObjectDoesNotExist("Missing attribute from data: {}".format(data))
 
-    metadata = {"channel_id": channel_id}
     exportchannel_task.delay(
         channel_id,
-        user_id=request.user.pk,
-        task_type='exportchannel',
-        is_progress_tracking=False,
-        metadata=metadata)
+        user_id=request.user.pk
+    )
 
     return HttpResponse(json.dumps({
         "success": True,
